+++ src/main/java/umc/cockple/demo/domain/exercise/controller/ExerciseController.java
@@ -15,6 +15,7 @@ import umc.cockple.demo.domain.exercise.enums.MyExerciseFilterType;
 import umc.cockple.demo.domain.exercise.enums.MyExerciseOrderType;
 import umc.cockple.demo.domain.exercise.enums.MyPartyExerciseOrderType;
 import umc.cockple.demo.domain.exercise.service.ExerciseCommandService;
+import umc.cockple.demo.domain.exercise.service.ExerciseGuestService;
 import umc.cockple.demo.domain.exercise.service.ExerciseParticipationService;
 import umc.cockple.demo.domain.exercise.service.ExerciseQueryService;
 import umc.cockple.demo.domain.party.enums.ActivityTime;
@@ -35,6 +36,7 @@ import java.util.List;
 public class ExerciseController {
 
     private final ExerciseCommandService exerciseCommandService;
+    private final ExerciseGuestService exerciseGuestService;
     private final ExerciseParticipationService exerciseParticipationService;
     private final ExerciseQueryService exerciseQueryService;
 
@@ -86,7 +88,7 @@ public class ExerciseController {
     ) {
         Long inviterId = SecurityUtil.getCurrentMemberId();
 
-        ExerciseGuestInviteDTO.Response response = exerciseCommandService.inviteGuest(
+        ExerciseGuestInviteDTO.Response response = exerciseGuestService.inviteGuest(
                 exerciseId, inviterId, request);
 
         return BaseResponse.success(CommonSuccessCode.CREATED, response);
