+++ src/main/java/umc/cockple/demo/domain/exercise/service/ExerciseParticipationService.java
@@ -58,7 +58,6 @@ public class ExerciseParticipationService {
         return exerciseConverter.toJoinResponse(savedMemberExercise, exercise);
     }
 
-
     public ExerciseCancelDTO.Response cancelParticipation(Long exerciseId, Long memberId) {
 
         log.info("운동 참여 취소 시작 - exerciseId: {}, memberId: {}", exerciseId, memberId);
